requirements.txt
Django==3.2.5
psycopg2==2.9.1
djangorestframework==3.12.4
celery==5.1.0